AI Drama
This is the Pathfinding Program now installed onto Unity. The system isn't fully functional yet (but works if you really want to look at it) as it is working opposite to how it should and is still a bit rough even though I realistically had it working and then decided to play around with it.